Поправлен кодстайл, дизайн текстовых полей и даты
